[unrated] Title: 과제 진행하기, Time: 14.41 ms, Memory: 74.5 MB -BaekjoonHub

+++ 프로그래머스/unrated/176962. 과제 진행하기/과제 진행하기.java
@@ -6,10 +6,10 @@ class Solution {
         // String[] answer = new String[plans.length];
         ArrayList<String> list = new ArrayList<>();
         Stack<String[]> stack = new Stack<>();
-        
+
         for(String[] plan : plans) {
             String[] time = plan[1].split(":");
-            plan[1] = Long.toString(Long.parseLong(time[0]) * 60 + Long.parseLong(time[1]));
+            plan[1] = Integer.toString(Integer.parseInt(time[0]) * 60 + Integer.parseInt(time[1]));
         }
         Arrays.sort(plans, (arr1, arr2) -> Integer.parseInt(arr1[1]) - Integer.parseInt(arr2[1]));
         for(int i = 0; i < plans.length; i++) {
@@ -19,22 +19,22 @@ class Solution {
                 continue;
             }
             String[] prevPlan = stack.peek();
-            long[] prev = {Long.parseLong(prevPlan[1]), Long.parseLong(prevPlan[2])};
-            long[] now = {Long.parseLong(plan[1]), Long.parseLong(plan[2])};
-            long remain = now[0] - prev[0];
+            int[] prev = {Integer.parseInt(prevPlan[1]), Integer.parseInt(prevPlan[2])};
+            int[] now = {Integer.parseInt(plan[1]), Integer.parseInt(plan[2])};
+            int remain = now[0] - prev[0];
             if(prev[1] > remain) {
-                prevPlan[1] = Long.toString(now[0]);
-                prevPlan[2] = Long.toString(prev[1] - remain);
+                prevPlan[1] = Integer.toString(now[0]);
+                prevPlan[2] = Integer.toString(prev[1] - remain);
                 stack.push(plan); 
                 continue;
             }
             while(remain > 0 && !stack.isEmpty()) {
                 prevPlan = stack.pop();
-                prev[1] = Long.parseLong(prevPlan[2]);
+                prev[1] = Integer.parseInt(prevPlan[2]);
                 if(prev[1] <= remain) list.add(prevPlan[0]);
                 else {
-                    prevPlan[1] = Long.toString(now[0]);
-                    prevPlan[2] = Long.toString(prev[1] - remain);
+                    prevPlan[1] = Integer.toString(now[0]);
+                    prevPlan[2] = Integer.toString(prev[1] - remain);
                     stack.push(prevPlan);
                 }
                 remain -= prev[1];
@@ -42,7 +42,7 @@ class Solution {
             stack.push(plan);
         }
         while(!stack.isEmpty()) list.add(stack.pop()[0]);
-        
+
         // return answer;
         return list;
     }
